requirements.txt
langchain==0.1.13
langchain_community==0.0.29
langchain_experimental==0.0.55
langchain_openai==0.0.8
pandas==2.2.1
python-dotenv==1.0.1
streamlit==1.32.2
